shell: physically split WasmShell into Headless base + view subclass (Phase 2b.3)
The contract types from Phase 2b.1-2 are now backed by real classes.
`wasm-shell-headless.ts` exports `WasmShellHeadless` — the worker-
safe base with just-bash, the VFS adapter, custom commands (git,
mount, supplemental), `.jsh` discovery + sync, and the
`executeCommand` / `executeScriptFile` primitives. `wasm-shell.ts`
shrinks to a `WasmShell` subclass that adds xterm mount, theme
sync, refit/resize, line editor, history, tab completion, Ctrl+C,
multi-line continuation, and inline media-preview rendering for
`imgcat`.

Inheritance shape:

  WasmShellHeadless implements HeadlessShellLike
    │
    └── WasmShell extends WasmShellHeadless
          ├── overrides runCommand to thread `execAbort.signal` (so
          │   terminal Ctrl+C cancels just-bash)
          └── overrides renderMediaPreview to draw to DOM (base
              throws "preview unavailable in headless mode")

Most fields moved to `protected` on the base so view methods can
read shared state (`bash`, `cwd`, `lastEnv`, `vfsAdapter`,
`scriptCatalog`, `ownsScriptCatalog`). `runCommand` is `protected`
on the base — `WasmShell.executeCommandInTerminal` and
`handleEnter` call it through the override that adds the
`execAbort` signal.

Re-exports preserved for backwards compat: existing tests import
`encodeForbiddenRequestHeaders` / `decodeForbiddenResponseHeaders`
/ `isTextContentType` / `WasmShell` from `wasm-shell.ts` — all
still resolve. New: `WasmShellHeadless` and `HeadlessShellLike`
also re-exported from `wasm-shell.ts` so callers have one canonical
import surface.

Verification:
  - typecheck (4 tsconfigs) clean
  - npm test: 3670 passed, 3 skipped (no changes — existing
    `tests/shell/wasm-shell.test.ts` exercises the same surface
    through the `WasmShell` subclass and goes through the
    inherited methods)
  - webapp + chrome-extension builds green

Worker bundle implication: the kernel worker still constructs
`WasmShell` (via scoop-context); inheritance carries no runtime
cost because xterm imports stay in `mount()` (lazy, never called
in worker). Phase 3+ may flip the worker to construct
`WasmShellHeadless` directly to make the type narrowing explicit;
this commit is a pure refactor and doesn't touch the construction
sites.
